requirements.txt
streamlit==1.38.0
opencv-python-headless
numpy
streamlit-drawable-canvas
pillow
moviepy==1.0.3
pymupdf
proglog

